requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.5
django-cors-headers==4.5.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
frozendict==2.4.6
html5lib==1.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
multitasking==0.0.12
numpy==2.3.2
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
yfinance==0.2.48
